// forms-web/src/app/generated/model/question.ts
export interface Question {
        /**
        * Unique document id auto generated
        */
        readonly id?: string;

        /**
        * The principal that created the entity containing the field.
        */
        readonly createdBy?: string;

        /**
        * The date and time the entity containing the field was created.
        */
        readonly createdAt?: Date;

        /**
        * The principal that recently modified the entity containing the field.
        */
        readonly updatedBy?: string;

        /**
        * The date the entity containing the field was recently modified.
        */
        readonly updatedAt?: Date;

        /**
        * The property owner id this object is associated with
        */
        propertyOwnerId?: string;

        /**
        * The language id this object is associated with
        */
        languageId?: string;

        type: QuestionType;

        /**
        * Question id
        */
        questionId: string;

        /**
        * Question Text.
        */
        text: string;

        /**
        * one or more question group like Client/Parent/Guardian
        */
        groups: Array<QuestionGroup>;

        /**
        * The options of the question for user to choose
        */
        options?: Array<QuestionOption>;

        possibleAnswers?: Array<string>;

        dataType: QADataType;

        validations?: Array<ValidationType>;

        additionalUserInputRequired?: boolean;

        additionalUserInputFieldLabel?: string;

        helpText?: string;

        order?: number;

        dependsOn?: Array<QuestionAllOfDependsOnInner>;

}